python-service/lightweight_3d_service.py
#!/usr/bin/env python3
"""
轻量级3D生成服务
支持多种轻量级模型，适合本地部署
"""

import os
import io
import base64
import logging
from typing import Optional, Dict, Any
from pathlib import Path

import torch
import numpy as np
from PIL import Image
from fastapi import FastAPI, HTTPException, UploadFile, File
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel

# 设置日志
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(title="轻量级3D生成服务", version="1.0.0")

# 添加CORS中间件
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://127.0.0.1:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 全局变量
models = {}

class GenerationRequest(BaseModel):
    prompt: str
    model_type: str = "shap-e"  # shap-e, dreamgaussian, instant3d
    quality: str = "medium"     # low, medium, high
    num_steps: int = 20

class ImageTo3DRequest(BaseModel):
    image_base64: str
    model_type: str = "zero123"  # zero123, pifu
    quality: str = "medium"

class GenerationResponse(BaseModel):
    success: bool
    task_id: str
    status: str
    result: Optional[Dict[str, Any]] = None
    error: Optional[str] = None

def load_lightweight_model(model_type: str):
    """加载轻量级模型"""
    if model_type in models:
        return models[model_type]
    
    try:
        if model_type == "shap-e":
            # 这里可以集成Shap-E
            logger.info("Shap-E模型加载中...")
            # models[model_type] = load_shap_e()
            
        elif model_type == "dreamgaussian":
            # 这里可以集成DreamGaussian
            logger.info("DreamGaussian模型加载中...")
            # models[model_type] = load_dreamgaussian()
            
        elif model_type == "instant3d":
            # 这里可以集成Instant3D
            logger.info("Instant3D模型加载中...")
            # models[model_type] = load_instant3d()
            
        elif model_type == "zero123":
            # 这里可以集成Zero-1-to-3
            logger.info("Zero-1-to-3模型加载中...")
            # models[model_type] = load_zero123()
            
        elif model_type == "pifu":
            # 这里可以集成PIFu
            logger.info("PIFu模型加载中...")
            # models[model_type] = load_pifu()
            
        else:
            raise ValueError(f"不支持的模型类型: {model_type}")
            
        return models.get(model_type)
        
    except Exception as e:
        logger.error(f"模型加载失败: {e}")
        return None

@app.get("/")
async def root():
    """健康检查"""
    return {
        "message": "轻量级3D生成服务运行中",
        "available_models": ["shap-e", "dreamgaussian", "instant3d", "zero123", "pifu"],
        "gpu_info": {
            "available": torch.cuda.is_available(),
            "device_name": torch.cuda.get_device_name(0) if torch.cuda.is_available() else "CPU",
            "memory_gb": round(torch.cuda.get_device_properties(0).total_memory / 1e9, 2) if torch.cuda.is_available() else 0
        }
    }

@app.post("/generate/text-to-3d", response_model=GenerationResponse)
async def text_to_3d(request: GenerationRequest):
    """文本生成3D模型"""
    task_id = f"text3d_{hash(request.prompt)}_{request.model_type}"
    
    try:
        # 加载模型
        model = load_lightweight_model(request.model_type)
        if model is None:
            raise HTTPException(status_code=500, detail="模型加载失败")
        
        # 模拟生成过程
        logger.info(f"开始文本生成3D: {request.prompt}")
        
        # 这里应该是实际的模型推理
        # result = model.generate(request.prompt, num_steps=request.num_steps)
        
        # 模拟返回
        return GenerationResponse(
            success=True,
            task_id=task_id,
            status="completed",
            result={
                "mesh_url": f"/files/{task_id}.glb",
                "thumbnail_url": f"/files/{task_id}_thumb.jpg",
                "format": "glb",
                "model_type": request.model_type,
                "generation_time": "45秒"
            }
        )
        
    except Exception as e:
        logger.error(f"文本生成3D失败: {e}")
        return GenerationResponse(
            success=False,
            task_id=task_id,
            status="failed",
            error=str(e)
        )

@app.post("/generate/image-to-3d", response_model=GenerationResponse)
async def image_to_3d(request: ImageTo3DRequest):
    """图片生成3D模型"""
    task_id = f"img3d_{hash(request.image_base64)}_{request.model_type}"
    
    try:
        # 解码图片
        image_data = base64.b64decode(request.image_base64)
        image = Image.open(io.BytesIO(image_data))
        
        # 加载模型
        model = load_lightweight_model(request.model_type)
        if model is None:
            raise HTTPException(status_code=500, detail="模型加载失败")
        
        # 模拟生成过程
        logger.info(f"开始图片生成3D，使用模型: {request.model_type}")
        
        # 这里应该是实际的模型推理
        # result = model.generate(image)
        
        # 模拟返回
        return GenerationResponse(
            success=True,
            task_id=task_id,
            status="completed",
            result={
                "mesh_url": f"/files/{task_id}.glb",
                "thumbnail_url": f"/files/{task_id}_thumb.jpg",
                "format": "glb",
                "model_type": request.model_type,
                "generation_time": "2分钟"
            }
        )
        
    except Exception as e:
        logger.error(f"图片生成3D失败: {e}")
        return GenerationResponse(
            success=False,
            task_id=task_id,
            status="failed",
            error=str(e)
        )

@app.get("/models/info")
async def get_models_info():
    """获取模型信息"""
    return {
        "text_to_3d_models": [
            {
                "name": "shap-e",
                "description": "OpenAI轻量级模型",
                "vram_required": "2-4GB",
                "generation_time": "30秒-2分钟",
                "quality": "中等"
            },
            {
                "name": "dreamgaussian", 
                "description": "高斯splatting技术",
                "vram_required": "3-6GB",
                "generation_time": "1-3分钟",
                "quality": "高"
            },
            {
                "name": "instant3d",
                "description": "极速生成模型",
                "vram_required": "3-4GB", 
                "generation_time": "10-30秒",
                "quality": "中等"
            }
        ],
        "image_to_3d_models": [
            {
                "name": "zero123",
                "description": "单图多视角生成",
                "vram_required": "4-6GB",
                "generation_time": "2-4分钟",
                "quality": "高"
            },
            {
                "name": "pifu",
                "description": "人物3D重建",
                "vram_required": "2-4GB",
                "generation_time": "1-2分钟", 
                "quality": "中等"
            }
        ]
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
