python-service/requirements.txt
# 轻量级3D生成服务依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3

# 深度学习框架
torch==2.1.0
torchvision==0.16.0

# 3D生成模型
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1

# 轻量级3D模型
shap-e==0.0.1

# 其他工具
requests==2.31.0
aiofiles==23.2.1